admin_api/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
requests==2.31.0
httpx==0.24.1
psycopg2-binary==2.9.8
python-dotenv==1.0.0
